Fixing issues with the new VG module
