+++ gdeployfeatures/vg/vg.py
@@ -3,9 +3,12 @@
 Add functions corresponding to each of the actions in the json file.
 The function should be named as follows <feature name>_<action_name>
 """
-from gdeploylib import defaults
+from gdeploylib import defaults, Helpers, Global
+
+helpers = Helpers()
 
 def vg_create(section_dict):
+    global helpers
     pvnames = get_pv_names(section_dict)
     vgnames = helpers.listify(section_dict['vgname'])
     section_dict = set_default_values(section_dict, pvnames, vgnames)
@@ -13,6 +16,7 @@ def vg_create(section_dict):
     return section_dict, defaults.VGCREATE_YML
 
 def vg_extend(section_dict):
+    global helpers
     pvnames = get_pv_names(section_dict)
     vgnames = helpers.listify(section_dict['vgname'])
     if len(vgnames) != 1:
@@ -23,11 +27,8 @@ def vg_extend(section_dict):
     return section_dict, defaults.VGEXTEND_YML
 
 def set_default_values(section_dict, pvnames, vgnames):
+    global helpers
     if len(pvnames) != len(vgnames):
-        if section_dict['one-to-one'] != 'no':
-            print "Error: number of vgname and pvname " \
-                    "does not match for one-to-one mapping."
-            helpers.cleanup_and_quit()
         if len(vgnames) > len(pvnames):
             print "Error: Insufficient number of values for pvname"
             helpers.cleanup_and_quit()
@@ -36,11 +37,11 @@ def set_default_values(section_dict, pvnames, vgnames):
                 print "Error: Provide 1 value for vgname or " \
                         "one for each pvname"
                 helpers.cleanup_and_quit()
-            if section_dict['one-to-one'] != 'no':
+            if section_dict['one-to-one'] == 'no':
                 vgnames *= len(pvnames)
             else:
                 vgs = []
-                for i in range(1, len(pvnames)):
+                for i in range(1, len(pvnames) + 1):
                     vgs.append(vgnames[0] + str(i))
                 vgnames = vgs
     return dictify_pv_vg_names(section_dict, pvnames, vgnames)
@@ -56,6 +57,7 @@ def dictify_pv_vg_names(section_dict, pvnames, vgnames):
     return section_dict
 
 def get_pv_names(section_dict):
+    global helpers
     pvnames = section_dict.get('pvname')
     if not pvnames:
         pv_section = Global.sections.get('pv')
@@ -65,5 +67,5 @@ def get_pv_names(section_dict):
             helpers.cleanup_and_quit()
         pvnames = pv_section['devices']
     pvnames = helpers.correct_brick_format(
-            helpers.listify(pv_names))
+            helpers.listify(pvnames))
     return pvnames
