+++ playbooks/vgcreate.yml
@@ -11,6 +11,7 @@
         diskcount="{{ diskcount }}"
         disktype="{{ disktype }}"
         stripesize="{{ stripesize }}"
+        dalign="{{ dalign }}"
     with_items: vgnames
     register: result
     failed_when: "result.rc != 0 and 'already exists' not in result.msg and 'is already in volume group' not in result.msg"
