+++ playbooks/vgextend.yml
@@ -7,7 +7,7 @@
   tasks:
   - name: Extend volume group
     vg: action=extend disks={{ item }}
-        vgname=vg
+        vgname="{{ vg }}"
     with_items: disk
     register: result
     failed_when: "result.rc != 0 and ('is already in volume group' not in result.msg  and 'not found' not in result.msg)"
